convert behandeling-van-agendapunt component to gts
