--- app/components/behandeling-van-agendapunt.hbs
@@ -1,207 +1,0 @@
-{{#if this.editable}}
-  <div class='au-c-meeting-chrome-editable'>
-    <div
-      class='au-u-flex au-u-flex--spaced-small au-u-flex--row au-u-flex--vertical-center'
-    >
-      <AuHeading id='behandeling-{{@behandeling.id}}' @level='3' @skin='4'>
-        {{limit-content @behandeling.onderwerp.titel 80}}
-      </AuHeading>
-      {{#if this.isInaugurationMeeting}}
-        <AuPill
-          @icon='alert-triangle'
-          @skin='warning'
-          class='au-u-hide-on-print'
-        >
-          {{t 'behandeling-van-agendapunten.sync-warning'}}
-        </AuPill>
-      {{/if}}
-    </div>
-    {{#unless @focusMode}}
-      <Common::MeetingSubSection
-        @title={{t 'behandeling-van-agendapunten.visibility'}}
-        class='au-u-hide-on-print'
-      >
-        <:body>
-          <div class='au-u-padding-tiny au-u-padding-bottom-none'>
-            {{#let (unique-id) as |id|}}
-              <AuLabel for={{id}}><Input
-                  @type='checkbox'
-                  @checked={{this.openbaar}}
-                  {{on 'input' (perform this.toggleOpenbaar)}}
-                  id={{id}}
-                />
-                &nbsp;{{t 'behandeling-van-agendapunten.openbaar-label'}}
-              </AuLabel>
-            {{/let}}
-            {{#if this.toggleOpenbaar.isRunning}}
-              <AuLoader @padding='small' @hideMessage={{true}}>{{t
-                  'application.loading'
-                }}</AuLoader>
-            {{/if}}
-          </div>
-        </:body>
-      </Common::MeetingSubSection>
-    {{/unless}}
-  </div>
-{{else}}
-  <div class='au-c-meeting-chrome-editable'>
-    <AuToolbar>
-      <AuHeading id='behandeling-{{@behandeling.id}}' @level='3' @skin='4'>
-        {{limit-content @behandeling.onderwerp.titel 80}}
-      </AuHeading>
-      {{#if this.published}}
-        <AuPill @skin='success'>
-          {{t 'manage-agenda-zitting-modal.ap-published-msg'}}
-        </AuPill>
-      {{/if}}
-    </AuToolbar>
-    {{#unless @focusMode}}
-      <Common::MeetingSubSection
-        @title={{t 'behandeling-van-agendapunten.visibility'}}
-      >
-        <:body>
-          <div class='au-u-padding-tiny'>
-            <AuPill @skin='border'>
-              {{#if this.openbaar}}
-                {{t 'behandeling-van-agendapunten.openbaar-msg'}}
-              {{else}}
-                {{t 'behandeling-van-agendapunten.geen-openbaar-msg'}}
-              {{/if}}
-            </AuPill>
-          </div>
-        </:body>
-      </Common::MeetingSubSection>
-    {{/unless}}
-  </div>
-
-{{/if}}
-{{#unless @focusMode}}
-  {{#if this.isLoading}}
-    <AuLoader @centered={{false}}>{{t
-        'participation-list.loading-title'
-      }}</AuLoader>
-  {{else}}
-    <ParticipationList
-      @chairman={{this.chairman}}
-      @defaultChairman={{this.defaultChairman}}
-      @secretary={{this.secretary}}
-      @defaultSecretary={{this.defaultSecretary}}
-      @participants={{this.participants}}
-      @defaultParticipants={{this.defaultParticipants}}
-      @absentees={{this.absentees}}
-      @defaultAbsentees={{this.defaultAbsentees}}
-      @possibleParticipants={{@possibleParticipants}}
-      @bestuursorgaan={{@bestuursorgaan}}
-      @onSave={{this.saveParticipants}}
-      @meeting={{@meeting}}
-      @modalTitle={{t 'generic.edit'}}
-      @readOnly={{or (not this.editable) (not this.canEditParticipants)}}
-      @readOnlyText={{unless
-        this.canEditParticipants
-        (t 'behandeling-van-agendapunten.cannot-edit-participants')
-      }}
-      @loading={{@loadingParticipants}}
-    />
-  {{/if}}
-  <Common::MeetingSubSection
-    @title={{t 'behandeling-van-agendapunten.voting-title'}}
-  >
-    <:body>
-      {{#if this.hasParticipants}}
-        <Treatment::Voting
-          @bestuursorgaan={{@bestuursorgaan}}
-          @behandeling={{@behandeling}}
-          @readOnly={{not this.editable}}
-        />
-      {{else}}
-        <div
-          class='au-u-margin-top au-u-margin-bottom au-u-flex au-u-flex--center'
-        >
-          <AuPill @skin='warning' @icon='alert-triangle' @iconAlignment='left'>
-            {{t 'behandeling-van-agendapunten.needs-participants'}}
-          </AuPill>
-        </div>
-      {{/if}}
-    </:body>
-    <:button>
-      {{#unless @readOnly}}
-        <div class='au-u-hide-on-print'>
-
-          <AuButton
-            {{on 'click' (perform this.addStandardVoting)}}
-            @icon='add'
-            @skin='secondary'
-            @iconAlignment='left'
-            @disabled={{not (and this.hasParticipants this.editable)}}
-            @loading={{this.addStandardVoting.isRunning}}
-            @loadingMessage={{t 'application.loading'}}
-          >
-            {{t 'generic.add'}}
-          </AuButton>
-          <AuButton
-            {{on 'click' (perform this.addCustomVoting)}}
-            @icon='add'
-            @skin='secondary'
-            @iconAlignment='left'
-            @disabled={{not (and this.hasParticipants this.editable)}}
-            @loading={{this.addCustomVoting.isRunning}}
-            @loadingMessage={{t 'application.loading'}}
-          >
-            {{t 'behandeling-van-agendapunten.add-custom-voting'}}
-          </AuButton>
-          {{#if this.editMode}}
-            <Treatment::Voting::Edit
-              @bestuursorgaan={{@bestuursorgaan}}
-              @onSave={{perform this.saveStemming}}
-              @onCancel={{this.onCancelEdit}}
-              @saving={{this.saveStemming.isRunning}}
-            />
-          {{/if}}
-        </div>
-      {{/unless}}
-    </:button>
-  </Common::MeetingSubSection>
-{{/unless}}
-<Common::MeetingSubSection
-  @title={{t 'behandeling-van-agendapunten.content-title'}}
->
-  <:body>
-    {{#if this.attachments}}
-      {{#if this.editable}}
-        <AuPill
-          @route='agendapoints.attachments'
-          @model={{@behandeling.documentContainer.id}}
-          @query={{hash returnToMeeting=@meeting.id}}
-          @skin='link'
-          @icon='attachment'
-          @iconAlignment='left'
-        >
-          {{this.attachments.length}}
-        </AuPill>
-      {{else}}
-        <AuPill @skin='border' @icon='attachment' @iconAlignment='left'>
-          {{this.attachments.length}}
-        </AuPill>
-      {{/if}}
-    {{/if}}
-    <div class='au-o-box au-o-box--small au-c-editor-preview say-content'>
-      {{@behandeling.documentContainer.currentVersion.htmlSafeContent}}
-    </div>
-  </:body>
-
-  <:button>
-    {{#if (and this.editable @behandeling.documentContainer)}}
-      <AuLink
-        @route='agendapoints.edit'
-        @model={{@behandeling.documentContainer.id}}
-        @query={{hash returnToMeeting=@meeting.id}}
-        @skin='button-secondary'
-        @icon='pencil'
-        @iconAlignment='left'
-        class='au-u-hide-on-print'
-      >
-        {{t 'generic.edit'}}
-      </AuLink>
-    {{/if}}
-  </:button>
-</Common::MeetingSubSection>
